image-to-excel-service/src/requirements.txt
pandas==2.1.4
openpyxl==3.1.2
boto3>=1.35.0
Pillow
